Add validation to pdf-block schema
When creating a faq question and response in the cms the user can add a pdf. However the pdf was not a required field so it was causing the front end not to respond to changes and breaking a build.
Added validation and an error message to ensure users always upload pdfs when using this component.